+++ src/app/api/db/editItemFromGroupContract/route.js
@@ -23,15 +23,3 @@ export async function PATCH(req) {
     return Response.json({ data: null, error: "Not authorized" });
   }
 }
-
-export async function POST(req) {
-  const session = await auth();
-
-  console.log(req);
-
-  if (!!session) {
-    return Response.json({ data: "POST HANDLED", error: null });
-  } else {
-    return Response.json({ data: null, error: "Not authorized" });
-  }
-}
